Added a check when the geometry process fails and goes to remove the failed content file that it was a file is in the local content folder as opposed to a core resource.
Summary: This prevents the odd situation when active development on geometry processor results in a failed import that deletes core resources.

Test Plan: Import a file exported in an unsupported format. It won't be in the content folder after so that it fails to process when you close and open the project again. Rename the bad file to a core resource name (like cube.fbx) and overwrite a core resource with it. Open the engine. When it fails the resource will not be deleted.

Reviewers: #zero_engine_developers, trevor.sundberg

Reviewed By: #zero_engine_developers, trevor.sundberg

Subscribers: trevor.sundberg, dane.curbow

Differential Revision: https://dev.zeroengine.io/D740

+++ Systems/Content/GeometryContent.cpp
@@ -568,10 +568,21 @@ void GeometryContent::BuildContent(BuildOptions& options)
         options.Failure = true;
         options.Message = String::Format("Failed to process Geometry. File '%s', See above Geometry Processor Error.", Filename.c_str());
 
-        // then remove the content items that were added because importing failed
-        String metaFile = BuildString(fullFilePath, ".meta");
-        DeleteFile(metaFile);
-        DeleteFile(fullFilePath);
+        // Get our projects content folder
+        ProjectSettings* projectSettings = Z::gEngine->GetProjectSettings();
+        String contentDirectory = FilePath::Normalize(projectSettings->GetContentFolder());
+
+        // If importing the item failed delete the content item but only if that item
+        // is in the projects local content folder since it was copied and will fail
+        // to be processed each time the project is opened until the user deletes the files
+        // manually. Otherwise a core resource failed and should not be deleted.
+        String fileDirectory = FilePath::Normalize(FilePath::GetDirectoryPath(fullFilePath));
+        if (fileDirectory == contentDirectory)
+        {
+          String metaFile = BuildString(fullFilePath, ".meta");
+          DeleteFile(metaFile);
+          DeleteFile(fullFilePath);
+        }
         return;
       }
       // let editor side importing know we have to load/reload the scene graph or textures
